Refactor sum_of_squares.c
This commit will refactor the computational algorithm for calculating
the sum of squares of any given natural number using "while loop"
iteration. The previous commit used "for loop" iteration for the same
purpose. The "while loop" interation was used to ascertain the
feasibility of its use in designing a computational algorithm for
calculating the sum of squares of any given natural number or positive
integer. The if-else conditional statements were used in order to
enforce that the user will only enter a natural number or positive
integer.