Add Conditional Statements to sum_of_squares.c
This commit will add if-else conditional statements to the computational
algorithm for calculating sum of squares of any given natural number
already created in the previous base branch with a "for loop" iteration.
The conditional statements were used to ensure that users, when
prompted, will input/enter only natural numbers or positive integers.